Extract out action in preparation for class splitting
